refactor: move files around and rename to premade-charts
