--- components/README.md
@@ -1,1 +1,0 @@
-# put the svelte components and js wrapper here
