+++ examples/README.md
@@ -1,14 +1,3 @@
 # Examples
 
-These are separate projects that can be run in the browser that are not part of the falcon package, but use it.
-
-## Development
-
-`cd` into the example of choice then
-
-```bash
-yarn # install packages
-yarn dev --open # to start the dev server and open the browser
-```
-
-changes to the example will be hot reloaded in the browser when file written.
+These will be examples that use the vanilla js, the premade components, for a variety of interesting data.
