--- examples/flights-in-browser/src/app.css
@@ -1,10 +1,0 @@
-:root {
-  --text-primary: grey;  
-
-  font-family: Inter, Avenir, Helvetica, Arial, sans-serif;
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 400;
-
-  color: var(--text-primary);
-}
