+++ package.json
@@ -7,7 +7,7 @@
   "license": "MIT",
   "scripts": {
     "test": "jest",
-    "dev": "cd examples/flights-in-browser && yarn && yarn dev --open",
+    "dev": "cd premade-charts && yarn && yarn dev --open",
     "build": "tsc"
   },
   "dependencies": {
