+++ premade-charts/README.md
@@ -1,0 +1,6 @@
+# put the svelte components and js wrapper here
+
+This is its own project with its own node modules because
+this may or may not be in a separate repo in the future.
+
+Under the hood this is a vite server for quick development see `package.json` for scripts.
